--- README.md
@@ -1,17 +1,0 @@
-# NAL - Números a Letras 
-[![Build Status](https://epconccs.visualstudio.com/NAL/_apis/build/status/NAL-Azure%20Web%20App%20for%20ASP.NET-CI?branchName=master)](https://epconccs.visualstudio.com/NAL/_build/latest?definitionId=1&branchName=master)
-[![codecov](https://codecov.io/gh/epconccs/numeros-a-letras/branch/master/graph/badge.svg)](https://codecov.io/gh/epconccs/numeros-a-letras)
-
-Este es un proyecto **ASP.NET Core Web API** para convertir números a letras en español.
-
-El API se encuentra publicada en el siguiente enlace -> https://nal.azurewebsites.net/
-
-Documentación de API -> https://nal.azurewebsites.net/swagger/
-
-El API está publicada en el marketplace de **RapidAPI** en el siguiente enlace -> https://rapidapi.com/epconmx/api/numeros-a-letras1
-
-
-
-Prueba -> https://nal.azurewebsites.net/api/NAL?num=256
-
-*Actualmente el proyecto únicamente acepta cifras de hasta 16 dígitos y acepta hasta 2 decimales.*
