--- nal.Test/NalControllerTest.cs
@@ -1,57 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using nal.Controllers;
-using Newtonsoft.Json;
-using Xunit;
-
-namespace nal.Test
-{
-    public class NalControllerTest
-    {
-
-        readonly NalController controller;
-
-        public NalControllerTest()
-        {
-            controller = new NalController();
-        }
-
-        [Fact]
-        public void Get_RespuestaOk()
-        {
-            var respuesta = controller.Get(256);
-
-            Assert.IsType<OkObjectResult>(respuesta);
-        }
-
-        [Fact]
-        public void Get_RespuestaBadRequest()
-        {
-            controller.ModelState.AddModelError("", "");
-
-            var respuesta = controller.Get(0);
-
-            Assert.IsType<BadRequestObjectResult>(respuesta);
-        }
-
-        [Fact]
-        public void Get_Numero_RespuestaLetra()
-        {
-            var numeroPrueba = new { letras = "DOSCIENTOS CINCUENTA Y SEIS" };
-
-            var respuesta = controller.Get(256) as OkObjectResult;
-
-            Assert.Equal(JsonConvert.SerializeObject(numeroPrueba), JsonConvert.SerializeObject(respuesta.Value));
-        }
-
-        [Fact]
-        public void Get_Numero_RespuestaFueraDeRango()
-        {
-            var fueraDeRango = new { letras = "Número fuera de rango." };
-
-            var respuesta = controller.Get(12345678901234567) as OkObjectResult;
-
-            Assert.Equal(JsonConvert.SerializeObject(fueraDeRango), JsonConvert.SerializeObject(respuesta.Value));
-        }
-
-    }
-}
