--- nal/Config/AppSettings.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace nal.Config
-{
-    public class AppSettings
-    {
-
-        public string RepoUrl { get; set; }
-
-    }
-}
