--- nal/appsettings.json
@@ -1,11 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Warning"
-    }
-  },
-  "AllowedHosts": "*",
-  "AppSettings": {
-    "RepoUrl": "https://github.com/epconccs/nal"
-  }
-}
