+++ numeros-a-letras-api.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.002.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "numeros-a-letras-api", "numeros-a-letras-api.csproj", "{C17F1E3E-857C-4749-AEF4-D7A84A1E2D78}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C17F1E3E-857C-4749-AEF4-D7A84A1E2D78}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C17F1E3E-857C-4749-AEF4-D7A84A1E2D78}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C17F1E3E-857C-4749-AEF4-D7A84A1E2D78}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C17F1E3E-857C-4749-AEF4-D7A84A1E2D78}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {9DBF8207-6473-4341-8403-506B15A525E8}
+	EndGlobalSection
+EndGlobal
